Implement Ajax for Screen Update
Implementing an ajax call to retrieve updates for the reception screen
rather than relying on a complete page refresh every 15 mins.
Prevents the screen attached to the raspberry pi from flashing due to
this page refresh and prevents a blank screen displaying if database or
web server is down when the update is called.

All around better solution.