Finish Ajax session update via backup record.
Allows user to update missing session details from the backup page and
updates the list of backup drives that can be selected if a composer is
added.

Also a tidy up of the layout within a the backup php pages. I was
a lot messier when I wrote this originally!